Stage 3/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reflective-assistant"
version = "0.1.0"
description = "Reflective Assistant - Journal & Todo AI agent (FastAPI)"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "psycopg2-binary",
    "groq",
    "alembic",
    "dateparser>=1.2.0",
    "asyncpg",
    "aiosqlite>=0.21.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]
